basic motion cares about the inclination of the robots body provide the system time for the sensor joint data added DCM pathses for the devices (not used now) added a debug thread for the naothmodule

+++ Framework/Platforms/Source/DCM/Tools/IPCData.h
@@ -37,7 +37,7 @@ namespace naoth
   class NaoSensorData
   {
   public:
-    unsigned int timeStamp;
+    unsigned long long timeStamp;
     float sensorsValue[numOfSensors];
     
     void get(SensorJointData& data) const;
